Delete old snapshots
Fix tx bounds for blocks without txs

Fix empty boundary block edge case

Handle newly registered accounts in snapshot

Rename and move to new file

setup cron job

Exclude retired pools

Check before running cache update

Missing THEN

Remove comment

cron run checks

Fix era identifier parsing

Solves issue reported in #1458

Fix snapshot boundaries

Add active connection count to get-metrics script (#1459)

Upper bounds for new accounts

Fix reward epoch limits

[gRest] Cast timestamps to int (#1460)

* Cast all timestamps to integers

* Missed cast in epoch_info

CNTools 10.0.1 (#1461)

* CNTools 10.0.1

Cold signing fix for pool registration / update. Last key was added twice when assemling witnesses due to not resetting tx_witness_files before assembling tx.

* cddl-format and mary era format on transaction building.

- Transactions now built using cddl-format to ensure that the formatting of transaction adheres the ledger specs.
- Default to mary era transaction building format for now.

* typo fix

Update delegation rules with regards to pool retire

Rename total_balance to amount

Add new Vasil data to tx_info (#1464)

* Add new Vasil data to tx_info

- Use jsonb instead of json due to 'could not identify an equality operator for type json' issue.
- collateral_outputs and reference_inputs added.
- collaterals renamed to collateral_inputs.
- datum_hash, inline_datum & reference_script added to collateral inputs/outputs, reference inputs & inputs/outputs json.

* Add actual cost_model instead of id reference

tx_utxos update (#1466)

Update tx_utxos to include optimizations from tx_info + asset fingerprint

Remove use of block_id <= .. and replace it with block_id = .. AND tx_count !=0

Performance optimizations

Further optimizations to reduce duplicated work

Add optimizations to both sections

Move last tx id in epoch to epoch_info_cache

Dont update current epoch

typo fix

Remove pool_info_cache refs

Temp table snapshot (#1473)

* Add capture last epoch's snaspshot function

* Delete old snapshots

* Fix tx bounds for blocks without txs

* Fix empty boundary block edge case

* Handle newly registered accounts in snapshot

* Fix uptime stat (#1457)

* Rename and move to new file

* setup cron job

* Exclude retired pools

* Check before running cache update

* Missing THEN

* Remove comment

* cron run checks

* Fix era identifier parsing

Solves issue reported in #1458

* Fix snapshot boundaries

* Add active connection count to get-metrics script (#1459)

* Upper bounds for new accounts

* Fix reward epoch limits

* [gRest] Cast timestamps to int (#1460)

* Cast all timestamps to integers

* Missed cast in epoch_info

* CNTools 10.0.1 (#1461)

* CNTools 10.0.1

Cold signing fix for pool registration / update. Last key was added twice when assemling witnesses due to not resetting tx_witness_files before assembling tx.

* cddl-format and mary era format on transaction building.

- Transactions now built using cddl-format to ensure that the formatting of transaction adheres the ledger specs.
- Default to mary era transaction building format for now.

* typo fix

* Update delegation rules with regards to pool retire

* Rename total_balance to amount

* Add new Vasil data to tx_info (#1464)

* Add new Vasil data to tx_info

- Use jsonb instead of json due to 'could not identify an equality operator for type json' issue.
- collateral_outputs and reference_inputs added.
- collaterals renamed to collateral_inputs.
- datum_hash, inline_datum & reference_script added to collateral inputs/outputs, reference inputs & inputs/outputs json.

* Add actual cost_model instead of id reference

* tx_utxos update (#1466)

Update tx_utxos to include optimizations from tx_info + asset fingerprint

* Add TEMP table and optimisations

* Add optimisations

* Update rules and optimisations

* Fix index statement

* fix stake_deregistration check

* Revert stake deregistration rule


add cron job removal of stake snapshot

Optimizations

1. Using i_last_tx_id from epoch_info_cache, removes ~1m from total query time.
2. account_delta_rewards, very slow query. 89M rows on mainnet for reward table and type nor spendable_epoch columns are indexed. ~1m40s on mainnet right now. Adding indexes on rewards table doesn't seem to help. Creating a temp table for rewards subset. This reduced execution time significantly. The fact that the same temp table can be reused for newly registered accounts multiplies this effect.
3. Removing unneeded variables.

Move tx_metalabels to use loose indexscan (#1474)

Remove unnecessary columns

Bump setup-grest

Further optimizations

Fix era identifier parsing

Solves issue reported in #1458

Add active connection count to get-metrics script (#1459)

[gRest] Cast timestamps to int (#1460)

* Cast all timestamps to integers

* Missed cast in epoch_info

CNTools 10.0.1 (#1461)

* CNTools 10.0.1

Cold signing fix for pool registration / update. Last key was added twice when assemling witnesses due to not resetting tx_witness_files before assembling tx.

* cddl-format and mary era format on transaction building.

- Transactions now built using cddl-format to ensure that the formatting of transaction adheres the ledger specs.
- Default to mary era transaction building format for now.

* typo fix

Add new Vasil data to tx_info (#1464)

* Add new Vasil data to tx_info

- Use jsonb instead of json due to 'could not identify an equality operator for type json' issue.
- collateral_outputs and reference_inputs added.
- collaterals renamed to collateral_inputs.
- datum_hash, inline_datum & reference_script added to collateral inputs/outputs, reference inputs & inputs/outputs json.

* Add actual cost_model instead of id reference

tx_utxos update (#1466)

Update tx_utxos to include optimizations from tx_info + asset fingerprint

+++ files/grest/rpc/address/address_txs.sql
@@ -3,7 +3,7 @@ CREATE FUNCTION grest.address_txs (_addresses text[], _after_block_height intege
     tx_hash text,
     epoch_no word31type,
     block_height word31type,
-    block_time numeric
+    block_time integer
   )
   LANGUAGE PLPGSQL
   AS $$
@@ -38,7 +38,7 @@ BEGIN
       DISTINCT(ENCODE(tx.hash, 'hex')) as tx_hash,
       block.epoch_no,
       block.block_no,
-      EXTRACT(epoch from block.time)
+      EXTRACT(epoch from block.time)::integer
     FROM
       public.tx
       INNER JOIN public.block ON block.id = tx.block_id
